+++ packages/jazz-tools/src/react/index.ts
@@ -8,6 +8,7 @@ export {
   useJazzContext,
   useAuthSecretStorage,
   useCoStateWithSelector,
+  useAccountWithSelector,
 } from "./hooks.js";
 
 export { createInviteLink, parseInviteLink } from "jazz-tools/browser";